requirements.txt
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.8.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-multipart>=0.0.6
pydantic>=2.0.0
numpy>=1.24.0
torch>=2.0.0